more progress. moving windows API layer to its own file
